// server/__generated__/type-resolvers.ts
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};

export type Comment = {
  __typename?: 'Comment';
  body: Scalars['String'];
  id: Scalars['ID'];
};

export type Item = {
  __typename?: 'Item';
  comments: Array<Comment>;
  id: Scalars['ID'];
  quantity: Scalars['String'];
  status: ItemStatus;
  title: Scalars['String'];
};

export enum ItemStatus {
  Active = 'ACTIVE',
  Inactive = 'INACTIVE',
  Pending = 'PENDING'
}